requirements.txt
python-telegram-bot==20.5
python-dotenv==1.0.0
openai==1.12.0
httpx==0.26.0
nest-asyncio==1.6.0
